==> Project Code/Portfolio/1 - NLP Preparation.ipynb <==
# Data Manipulation
#-----------------------------
import pandas as pd


# Natural Laguage Processing
#-----------------------------
import nltk

# Stopwords
from nltk.corpus import stopwords
nltk.download('stopwords')
nltk.download('wordnet')

# Punctuation
from nltk.tokenize import wordpunct_tokenize

# Stemmers and Lemmatizer
from nltk.stem.porter import PorterStemmer
from nltk.stem.lancaster import LancasterStemmer
from nltk.stem import WordNetLemmatizer, SnowballStemmer


# Other packages
#-----------------------------

import string
import re
import random
from random import sample

# Import Dataset

artist = pd.read_csv('../Data/Original Data/artists-data.csv')
lyrics = pd.read_csv('../Data/Original Data/lyrics-data.csv')

## Artist dataset 

artist.head()

print(f'The artist dataframe has {artist.shape[0]} rows and {artist.shape[1]} columns.')

artist.Genre.value_counts()

We only keep the Rock, Hip Hop and Pop songs, as most of them have lyrics in English.
The other genres consist of songs in Portuguese.

artist = artist[artist.Genre.isin(['Rock','Hip Hop','Pop'])]

artist.shape

We still have 65% of the original dataframe.

We now check if there are any null values.

artist.isnull().sum()

The null values in Genres are not an issue, as they indicate if artists belong to multiple genre, outside of their main one.
We will not use this data for our models.

genre_per_artist = artist.groupby(['Artist'])[['Genre']].count()
genre_per_artist.value_counts()

Some artists belong two 2 or more different (main) music genre. We will not keep them, to avoid any ambiguity when trying to classify the lyrics.

artist_1_genre = genre_per_artist[genre_per_artist.Genre == 1].index.tolist()

## Lyrics dataset

lyrics.head()

print(f'The lyrics dataframe has {lyrics.shape[0]} rows and {lyrics.shape[1]} columns.')

We keep the songs in English only.

lyrics = lyrics[lyrics.Idiom == 'ENGLISH']

While we lose about 50% of the dataset by doing this, we still have 114,723 songs in the dataset, which will be sufficiant for modelling.

We now check if there are any null values.

lyrics.isnull().sum()

## Merging the datasets

df_lyrics = artist.merge(lyrics, left_on='Link', right_on='ALink', how = 'inner')

We now drop the duplicate rows.

df_lyrics = df_lyrics.drop_duplicates()

We keep the songs where the artists only belong to 1 principal music genre.

df_lyrics = df_lyrics[df_lyrics.Artist.isin(artist_1_genre)]
df_lyrics = df_lyrics.reset_index(drop=True)

print(f'There are {len(df_lyrics.Artist.unique())} different artists.')

Genre_dict = {'Rock':0, 'Hip Hop':1, 'Pop':2}
df_lyrics['Genre'] = df_lyrics['Genre'].map(Genre_dict)

df_lyrics.head(2)

For convenience and readability we will only keep a few columns, as we will not be using the others ones.

df_lyrics = df_lyrics[['SName', 'Lyric','Artist', 'Genre']].copy()

print(f'We now have a dataframe made of {df_lyrics.shape[0]} different songs, ready for cleaning.')

We now export the dataframe, to save a copy of it.

df_lyrics.to_csv('../Data/Processed Data/music_1_genre.csv', index = False)

#df_lyrics = pd.read_csv('../Data/Processed Data/music_1_genre.csv')

## Radom lyrics sample

We have a look at a few songs.

random.seed(10)
dict = {0:'Rock', 1:'Hip Hop', 2:'Pop'}

for item in sample(range(len(df_lyrics)), 3):
    print(df_lyrics['Lyric'][item], "\n")
    print("Genre: {}".format(dict.get(df_lyrics['Genre'][item])))
    print("\n~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~\n")

# Data Preparation

## Stemming

Before cleaning the data, we explore what kind of stemmers are available, that is:
 - Porter Stemmer
 - Lancaster Stemmer
 - Snowball Stemmer
 
Because of the computational requirement for lemmatizing, we will only stem the data.

p_stemmer = PorterStemmer()
l_stemmer = LancasterStemmer()
s_stemmer = SnowballStemmer(language='english')

lemma = nltk.wordnet.WordNetLemmatizer()

my_str = ['apparatus', 'applicable', 'apply', 'apple', 'appletown', 
          'AppleTown', 'apples', 'apply', 'app', 'application', 'applied', 'applies', ]

def StemTest(my_str=my_str):
    ''' Testing various words on a selection of stemmers and lemmatizers '''
    
    assert(type(my_str)==list), "Please input your selection of words as a LIST!"
    my_str_1 = [p_stemmer.stem(x) for x in my_str] ## Use the Porter stemmer
    my_str_2 = [l_stemmer.stem(x) for x in my_str] ## Use the Lancaster stemmer
    my_str_3 = [s_stemmer.stem(x) for x in my_str] ## Use the Snowball stemmer
    my_str_4 = [lemma.lemmatize(x) for x in my_str]
    quick_dict = {1:'Porter: ', 2:'Lancaster: ', 3:"Snowball: ", 4:'Lemmatizer:'}

    ## Output our results from all stemmers, together with the original string
    print("Original: ", my_str, "\n")
    for i in range(1,5):
        print(quick_dict[i], locals()['my_str_{}'.format(i)], sep=" ")

We play around with the different stemmers/lemmatizer.

StemTest()

StemTest(['try', 'tryy', 'tryyy', 'tryyyy', 'tryin\''])

None of the stemmers or lemmatizer seem to be dealing well with words ending by multiple instances of the same letter (commonly found in song lyrics).

StemTest(['yeeeaaaahhhhh'])

We make a similar observation when the repetition of the letter is within a word.

## Data cleaning using Regex

To solve the letter repetition within words issue, we will create a RegEx.

def regex_clean(txt, regex, sub=' '):
    
    '''This function takes in a string and a regular expression.
    It will clean the string by removing any match from the regular expression.
    
    The output is the cleaned string'''
        
    return " ".join(re.sub(regex, sub, txt).split()) ## Substitute the desired regex with nothing,
                                                    ## then bring the sentence back together

### Repeated instances of letters

This snippet will remove letters that repeat themselves 3 times or more, and keep only a single instance.

test_items = ['ooooooh', 'yeeeaaaaahhhhh', 'good', 'hello', 'yassss!!!']
test_clean = [regex_clean(word, r'(\w)\1{2,}', r'\1') for word in test_items]

test_clean

### Instances of trailing apostrophes

For example: Trying -> tryin'
This kind of spelling is quite common within song lyrics.

test_items = ['tryin\' ', 'handlin\' ', 'handlin\'.', 'handling\'']
test_clean = [regex_clean(word, r"'\s|'\.", r'g') for word in test_items]

test_clean

### Instances of square brackets

We need to remove square brackets from the lyrics, as well as what is inside the brackets.

For example: \[Chorus 1\]

test_items = ['I once ate an elephant [It was delicious!]', '[Chorus 2] I will live on!', 
             'There [mouse] will be [rat] justice.']
test_clean = [regex_clean(word, r'(\[.*?\])', r'') for word in test_items]

test_clean

### Tracking

We will now keep track of how many instance of trailing ' and letter repetition there are in every song.
We already suspect that there should be a correlation between the number of ' and the Hip Hop genre.

# Cleaning

lyric_multiple_letter = [regex_clean(word, r'(\w)\1{2,}', r'\1') for word in df_lyrics.Lyric] # Removing all trailing '
lyric_clean_trail = [regex_clean(word, r"'\s|'\.", r'g') for word in df_lyrics.Lyric] # Removing triple letters or more. 

# Tokenising
lyric = [nltk.wordpunct_tokenize(sentence) for sentence in df_lyrics.Lyric] # We will use this list for comparison

lyric_multiple_letter = [nltk.wordpunct_tokenize(sentence) for sentence in lyric_multiple_letter]
lyric_clean_trail = [nltk.wordpunct_tokenize(sentence) for sentence in lyric_clean_trail]

#### Repeated instances of letters

To count how many time letters were repeated three times or more, we use the Hamming distance.

from scipy.spatial.distance import hamming

Example of how the Hamming distance works. 
The more similar the words are, the closer to 0 the score is.

hamming(['apple', 'apple', 'appl'], ['apple', 'apple', 'apple'])

hamming(['apple', 'apple', 'apple'], ['apple', 'apple', 'apple'])

We now define a function to calculate the Hamming distance.

def hamham(my_lst):
    '''This function calculates the Hamming distance between the lyrics and their tokenized version.
    Input = list with cleaned lyrics
    
    Output = list containing the hamming distance between the original lyrics and the cleaned lyrics'''
    
    new_lst = []
    j=0
    for item in my_lst:
        new_lst.append(hamming(item,lyric[j]))
        j+=1
    
    return new_lst

We now store the results in the main dataframe.

df_lyrics['multiple_letter'] = hamham(lyric_multiple_letter)

We look at the obtained values.

df_lyrics['multiple_letter'].describe()

We notice that the maximum is 0.27, which is not very high and the mean is very close to zero.
We can assume that this feature will not have too much impact on the modelling.

#### Instances of letters

Trails = [len(lyric[i])-len(lyric_clean_trail[i]) for i in range(len(lyric))]

df_lyrics['Trails'] = Trails

df_lyrics.Trails.value_counts()

There are some negative values, we set them to 0.

df_lyrics.loc[df_lyrics['Trails'] < 0, 'Trails'] = 0

# Data Preparation

stpwrds = stopwords.words('english') # Import the list of stopwords.
string.punctuation # Import the list of punctuation.
stpwrds.extend(string.punctuation) # Merge the 2 lists together.

#  Add chorus and verse to the list, as they are not part of the lyrics
stpwrds.extend(['chorus', 'verse', 'verses', 'choruses'])
                                        

### Function Declaration

def prep_data(song, stem='s'):
    ''' This function takes in a string, clean it using regular expressions, tokenize it, remove
    digits and any word shorter than 2 characters.
    
    It returned the clean and tokenized version of the string'''
    assert(stem=='s' or stem=='l' or stem=='p'), '''Input a correct stemming parameter and try again.
    The only accepted types are s for Snowball, l for Lancaster or p for Porter. Default is Snowball.'''
    
    song = song.lower() ## Transform the sentence into lowercase

    song = regex_clean(song, r'(\w)\1{2,}', r'\1') # Removing all trailing '
    song = regex_clean(song, r"'\s|'\.", r'g') # Removing triple letters or more.
    song = regex_clean(song, r'(\[.*?\])', r'') # Removing any characeters inside brackets (including brackets)
    song = regex_clean(song, r'(\W){2,}', r'\1') # Removing trailing white space
    
    ## Tokenize 
    song = nltk.wordpunct_tokenize(song) ## tokenize the string

    ## Post Token Cleaning - Stuff that applies to a list 
        
    song = [word for word in song if word not in stpwrds] # Eliminate all extended stopwords from among our tokens
    song = [globals()['{}_stemmer'.format(stem)].stem(word) for word in song] # Apply the chosen stemmer
    song = [word for word in song if not word.isdigit()] # Remove all digits
    song = [word for word in song if len(word)>2] # Remove words shorter than 2 characters, to avoid 'de, tg, ll' etc.


    return song

Function test:

prep_data("I am gonna go so hardn't I won't back down dwag 555!.?  ??any word", 'p')

### Cleaning

We now clean the lyrics and add them to the main dataframe.

lyrics = [" ".join(prep_data(x)) for x in df_lyrics['Lyric']]

df_lyrics['lyrics_clean'] = lyrics

Null values verification:

df_lyrics.isnull().sum()

### Extra features

We create a list containing the 10 most common words for each genre, after removing the 200 most popular words from each of the other genres.

Hop = set(['nigga', 'shit', 'bitch', 'fuck', 'hit', 'gon', 'niggaz', 'with', 'bout', 'hoe', 'cuz'])
Rock = set(['blue', 'dead', 'line', 'woman', 'child', 'cold', 'town', 'year',
       'help', 'alive', 'behind'])
Pop = set(['kiss', 'touch', 'christmas', 'hurt', 'beautiful', 'matter', 'somebody', 'trying', 'else', 'knew', 'forget'])

We will use set manipulations to check if one of these words appear in a song.

set_lyric = [set(song) for song in lyric] # Set of the tokenized lyrics

# These three lists will track if a word from the 10 most common words for each genre is present in the lyrics.
Rock_lst = [1 if len(song&Rock)>0 else 0 for song in set_lyric]
Pop_lst = [1 if len(song&Pop)>0 else 0 for song in set_lyric]
Hop_lst = [1 if len(song&Hop)>0 else 0 for song in set_lyric]

We now add these lists to the main dataframe, in 3 separate columns.

df_lyrics = pd.concat([df_lyrics, pd.Series(Rock_lst), pd.Series(Pop_lst), pd.Series(Hop_lst)], axis=1)
df_lyrics.rename(columns={0: 'Rock', 1:'Pop', 2:'Hip_hop'}, inplace=True)
df_lyrics.head()

df_lyrics.isnull().sum()

df_lyrics.iloc[7509]

Now that the dataframe is cleaned and ready for modelling, we export it.

df_lyrics.to_csv('../Data/Processed Data/clean_lyrics.csv', index = False)
